Pretty good tokenizer, incomplete and rough sketch of tokenizer
The tokenizer has no known bugs (but probably still has some)

The tokenizer doesn't have support for requirement sets yet, nor entitlements (but they're basically identical to info, so intentionally holding off on that for the moment). The organization of things in the tokenizer is pretty rough particularly in terms of protocols implemented, naming, etc. Lots of refactoring to be done before there's a nice API.

Decent number of tests so far and they're all passing. Definitely needs more testing to match Apple's compiler behavior which isn't entirely the same as what they documented.